moved pages out of components folder
-moved pages out of components folder
-added rating error messages
-added StarRating component to ComponentTesting.js
