--- frontend/src/components/Login/styles.module.css
@@ -1,113 +1,0 @@
-.login_container {
-	background-image: url(./background.png);
-	height: 100%;
-	background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-	width: 100%;
-	min-height: 100vh;
-	background-color: #f5f5f5;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
-.login_form_container {
-	width: 900px;
-	height: 500px;
-	display: flex;
-	border-radius: 10px;
-	box-shadow: 0px 3px 3px -2px rgb(0 0 0 / 20%),
-		0px 3px 4px 0px rgb(0 0 0 / 14%), 0px 1px 8px 0px rgb(0 0 0 / 12%);
-}
-
-.left {
-	flex: 2;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	background-color: white;
-	border-top-left-radius: 10px;
-	border-bottom-left-radius: 10px;
-}
-
-.form_container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-.form_container h1 {
-	font-size: 40px;
-	margin-bottom: 10px;
-}
-
-.input {
-	outline: none;
-	border: none;
-	width: 370px;
-	padding: 15px;
-	border-radius: 10px;
-	background-color: #edf5f3;
-	margin: 5px 0;
-	font-size: 14px;
-}
-
-.error_msg {
-	width: 370px;
-	padding: 15px;
-	margin: 5px 0;
-	font-size: 14px;
-	background-color: #f34646;
-	color: white;
-	border-radius: 5px;
-	text-align: center;
-}
-
-.right {
-	flex: 1;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	background-color: #060c0b;
-	border-top-right-radius: 10px;
-	border-bottom-right-radius: 10px;
-}
-
-.right h1 {
-	margin-top: 0;
-	color: white;
-	font-size: 40px;
-	align-self: center;
-}
-
-.white_btn,
-.green_btn {
-	border: none;
-	outline: none;
-	padding: 12px 0;
-	background-color: white;
-	margin: 10px;
-	border-radius: 20px;
-	width: 180px;
-	font-weight: bold;
-	font-size: 15px;
-	cursor: pointer;
-}
-
-.black_btn {
-	border-radius: 25;
-	background-color: #040e0c;
-	color: white;
-	margin: 10px;
-	border: none;
-	outline: none;
-	border-radius: 20px;
-	width: 100px;
-	height: 40px;
-	font-weight: bold;
-	font-size: 14px;
-	cursor: pointer;
-}
