--- frontend/src/components/Profile/styles.module.css
@@ -1,81 +1,0 @@
-*{
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  }
-  
-  .container {
-	margin: auto;
-	width: 100%;
-	max-width: 600px;
-  }
-  
-  .errorMsg {
-	color: red;
-	font-size: 24px;
-	margin: 3px 0;
-  }
-  
-  .profilePic {
-	width: 100%;
-	height: 900px;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	text-align: center;
-	color: #fff;
-  }
-  
-  .profilePic input[type="file"] {
-	display: none;
-  }
-  
-  .profilePic .customFileUpload {
-	color: #f5f9f9;
-	font-size: 22px;
-	font-weight: 500;
-	padding: 6px 12px;
-	cursor: pointer;
-  }
-  
-  button {
-	border: none;
-	outline: none;
-	width: 100%;
-	margin: 10px 0 0 0;
-	padding: 14px 10px;
-	background: #55d6d6;
-	color: #0a0a0a;
-	font-size: 17px;
-	font-weight: 600;
-	cursor: pointer;
-  }
-  .navbar {
-	width: 100%;
-	height: 70px;
-	background-color: #050a09;
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.navbar h1 {
-	color: white;
-	font-size: 25px;
-	margin-left: 20px;
-}
-
-.white_btn {
-	border: none;
-	outline: none;
-	padding: 12px 0;
-	background-color: white;
-	border-radius: 20px;
-	width: 120px;
-	font-weight: bold;
-	font-size: 14px;
-	cursor: pointer;
-	margin-right: 20px;
-}
