+++ frontend/src/components/StarRating/StarRating.module.css
@@ -1,0 +1,3 @@
+input[type="radio"]{
+  display: none
+}
