+++ frontend/src/styles/Profile.module.css
@@ -1,0 +1,60 @@
+.container {
+  margin: auto;
+  width: 100%;
+  max-width: 600px;
+}
+
+.errorMsg {
+  color: red;
+  font-size: 24px;
+  margin: 3px 0;
+}
+
+.profilePic {
+  width: 100%;
+  height: 900px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  color: #fff;
+}
+
+.profilePic input[type="file"] {
+  display: none;
+}
+
+.profilePic .customFileUpload {
+  color: #f5f9f9;
+  font-size: 22px;
+  font-weight: 500;
+  padding: 6px 12px;
+  cursor: pointer;
+}
+
+button {
+  border: none;
+  outline: none;
+  width: 100%;
+  margin: 10px 0 0 0;
+  padding: 14px 10px;
+  background: #55d6d6;
+  color: #0a0a0a;
+  font-size: 17px;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.white_btn {
+  border: none;
+  outline: none;
+  padding: 12px 0;
+  background-color: white;
+  border-radius: 20px;
+  width: 120px;
+  font-weight: bold;
+  font-size: 14px;
+  cursor: pointer;
+  margin-right: 20px;
+}
