--- frontend/src/styles/StarRating.css
@@ -1,3 +1,0 @@
-input[type="radio"]{
-  display: none
-}
