--- frontend/src/styles/calculateCost.css
@@ -1,16 +1,0 @@
-#chooseRide > li {
-    background-color: lightgrey;
-}
-
-li > p {
-    padding: 10px;
-    margin: 5px;
-}
-
-#caption {
-    font-size: 0.8em;
-}
-
-#title {
-    font-weight: bold;
-}
